+++ flatlaf-core/src/test/java/com/formdev/flatlaf/ui/TestFlatStyleableInfo.java
@@ -539,6 +539,7 @@ public class TestFlatStyleableInfo
 			"maximumThumbSize", Dimension.class,
 			"allowsAbsolutePositioning", boolean.class,
 
+			"minimumButtonSize", Dimension.class,
 			"trackInsets", Insets.class,
 			"thumbInsets", Insets.class,
 			"trackArc", int.class,